+++ CHANGELOG.md
@@ -52,3 +52,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Emails to BRIDGHEAD_ADMINS_WHO_HAVE_NOT_ACCEPTED_NOR_REJECTED_THE_PROJECT
 - FrontendSiteModule repeatable
 - Action names
+- Module names